Use PR data from the store
After some spelunking, and looking for `createStore` calls, I found a module `bitbucket/internal/feature/pull-request/store/pull-request-store` that exports the pull request store object which contains everything we need. This is more robust as we don't have to parse it from the DOM and it copies your PR message exactly.